Added Move to Album Features

+++ app/src/main/java/com/app/galleryx/gallery/albums/detail/ui/AlbumDetailViewModel.kt
@@ -1,17 +1,17 @@
 /*
- *   Copyright 2020–2026 Leon Latsch
+ * Copyright 2020–2026 Leon Latsch
  *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *        http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.app.galleryx.gallery.albums.detail.ui
@@ -131,14 +131,18 @@ class AlbumDetailViewModel @AssistedInject constructor(
                 }
             }
 
-            is AlbumDetailUiEvent.RemoveFromAlbum -> {
+            // CHANGED: Logic to move photos to another album
+            is AlbumDetailUiEvent.MoveToAlbum -> {
                 viewModelScope.launch {
-                    albumsRepository.unlink(event.items, albumFlow.value.uuid)
-                    navEventsChannel.trySend(
-                        ShowToast(
-                            resources.getString(R.string.common_ok)
+                    if (event.targetAlbumUuid != albumFlow.value.uuid) {
+                        albumsRepository.link(event.items, event.targetAlbumUuid)
+                        albumsRepository.unlink(event.items, albumFlow.value.uuid)
+                        navEventsChannel.trySend(
+                            ShowToast(
+                                resources.getString(R.string.gallery_albums_moved)
+                            )
                         )
-                    )
+                    }
                 }
             }
